+++ src/components/UI/Header/Header.module.css
@@ -1,106 +1,11 @@
-header {
+.header {
   position: fixed;
   top: -1rem;
   left: 0;
-  width: 100%;
   z-index: 100;
-}
-
-.main-nav {
   background-color: whitesmoke;
   width: 100%;
   height: 5rem;
-  display: flex;
-  align-items: center;
-  /* border-radius: 16px; */
   border: 2px solid black;
   box-shadow: 0 2px 10px rgba(0, 0, 0, 0.4);
-}
-
-.main-nav ul {
-  margin: 0;
-  padding: 0 1rem;
-  list-style: none;
-  display: flex;
-  justify-content: flex-end;
-  width: 100%;
-  position: relative;
-  top: 0.5rem;
-}
-
-.main-nav li {
-  margin: 0 0 0 1.5rem;
-}
-
-.main-nav a {
-  text-decoration: none;
-  color: black;
-  font-size: 1.1rem;
-  font-weight: bold;
-  transition: transform 200ms ease-out;
-}
-
-.main-nav li:hover,
-.main-nav li:active {
-  transform: translateX(-0.1rem) translateY(-0.1rem);
-  text-shadow: 2px 2px 2px rgba(0, 0, 0, 0.4);
-}
-
-.usrCircle {
-  scale: 2.3;
-}
-
-.user-d {
-  display: flex;
-  width: 9rem;
-  justify-content: space-between;
-  margin: 0 1rem;
-  cursor: pointer;
-  height: 100%;
-  align-items: center;
-}
-
-.r-arrow {
-  rotate: 270deg;
-  transition: rotate 500ms ease-out;
-}
-
-.r-arrow-clicked {
-  rotate: 450deg;
-}
-
-.logout {
-  display: none;
-  position: relative;
-  right: 5rem;
-  animation: lout 500ms ease-in;
-  overflow: hidden;
-}
-
-.logout-clicked {
-  display: flex;
-  width: 5rem;
-  justify-content: space-between;
-  right: 0;
-}
-
-@keyframes lout {
-  0% {
-    right: 5rem;
-  }
-  20% {
-    right: 4rem;
-  }
-  40% {
-    right: 3rem;
-  }
-  60% {
-    right: 2rem;
-  }
-  80% {
-    right: 1rem;
-  }
-  100% {
-    right: 0;
-  }
 }
