+++ src/components/UI/Header/MobileNav.module.css
@@ -1,0 +1,7 @@
+.mobile-nav {
+    display: flex;
+    align-items: center;
+    height: 100%;
+    justify-content: right;
+    margin: 0 0.5rem;
+}
