+++ src/components/UI/StackedBtn/StackedBtn.module.css
@@ -1,0 +1,29 @@
+.stacked-btn {
+  display: flex;
+  height: 80%;
+  min-width: 3rem;
+  background-color: whitesmoke;
+  border: 1px solid black;
+  border-radius: 7px;
+  cursor: pointer;
+  position: relative;
+  top: 0.4rem;
+  flex-direction: column;
+  justify-content: space-around;
+  padding: 0.3rem;
+  margin-left: 0.5rem;
+}
+
+.stack {
+  display: block;
+  width: 100%;
+  height: 20%;
+  background-color: black;
+  border-radius: 12px;
+}
+
+@media (min-width: 40rem) {
+  .stacked-btn {
+    display: none;
+  }
+}
