+++ src/components/UI/StackedBtn/StackedBtn.tsx
@@ -1,0 +1,13 @@
+import classes from './StackedBtn.module.css'
+
+const StackedBtn = () => {
+  return (
+    <button className={classes['stacked-btn']}>
+      <span className={classes.stack}></span>
+      <span className={classes.stack}></span>
+      <span className={classes.stack}></span>
+    </button>
+  )
+}
+
+export default StackedBtn
